requirements.txt
aiogram==3.15.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
aiosqlite==0.20.0
pydantic-settings==2.6.1
httpx==0.28.1
apscheduler==3.10.4
python-dotenv==1.0.1
